Fully decouple read and write sides.
This is achieved by splitting read and write into separate threads, with
a protected buffer in between. At present there's no constraint on the
size of the buffer.

Testing with the demo code at
https://www.idontplaydarts.com/static/bashserver.zip shows that this
does indeed protect against the bash detection method described at
https://www.idontplaydarts.com/2016/04/detecting-curl-pipe-bash-server-side/

Don't pipe code downloaded from the Internet directly into bash! That's
stupid. But if you want to decouple the sending process behaviour from the
receiving process behaviour in a shell command this code does that.