Implement a constrained buffer size.
This puts a limit on the number of entries in the buffer based on a
command line argument. When the buffer fills up, the read thread spins
(via yield_now()) waiting for the write thread to partly empty the
buffer.

There's lots of room for improving this, including things like having
the read thread wait for the buffer to have more than 1 slot available
(this would allow the read thread to operate in a more bursty manner
rather than doing a little bit and then spinning until it can do another
little bit), and using a self-tuning sleep period rather than yield in
order to allow the read thread to match its read speed to the write
speed of the write thread (though this also has potential to allow the
kind of information leak that this code was trying to protect against).
That said, synchronising based on the rate at which the write thread
frees up a slot in the buffer has almost exactly the same problem . . .